feat: create JoinList component
